Clean up and normalize Sekiro options
This is mostly just making sure everything uses the same wording/formatting as everything else and clarifying some ambiguities. However, I did get rid of the "get 3 items from hags" entirely because I think it's *too* easy—hags will give you an item if you use any mibu balloon near them, and it's not at all challenging to either find the balloons or make it to a hag, so this was essentially a free space.