Tweak miniboss challenges
This starts all "kill X [type]s" to at 2, since most randomizer runs will quickly encounter at least one of most types of minibosses. The only exception is Headless, because the randomizer defaults to making them optional and non-random and because killing even one is such a chore.

It also adds an option for killing Drunkard-type minibosses, and expands the Centipede option to cover all six miniboss types for which exactly two exist.

As always, this is just intended as a suggestion—feel free to keep either half of this or otherwise tweak as you desire.